+++ components/switcher/jobs/add/JobAddStepTwo.tsx
@@ -1,41 +1,24 @@
-import React, { useState } from "react"
+import React from "react"
 import Progress from "../../../common/Progress"
-import InputComponent from "../../../common/InputComponent"
 import {
 	FormActions,
 	PrimaryButton,
 	SecondaryButton,
-	FormInputContainer,
 	FormHeader
 } from "../../../styles/GlobalComponents"
-import { JobAddInterface } from "../../../../entities/interface/job"
 import { useRouter } from "next/router"
-import { offerCalculator } from "../../../utilities/costCalculater"
+import styled from "styled-components"
+import JobFormStepTwo from "../../../common/JobFormStepTwo"
 
-const JobAddStepTwo = (props: JobAddInterface) => {
+const FormActionsCustom = styled(FormActions)`
+	padding: 0 2.6rem;
+	margin: 1.2rem 0 3rem 0;
+`
+
+const JobAddStepTwo = () => {
 	const router = useRouter()
-	const { details, setDetails } = props
-	const [stepTwoDetails, setStepTwoDetails] = useState({
-		...details,
-		weight: String(details.weight),
-		waiting_time: String(details.waiting_time),
-		offer_price: String(details.offer_price)
-	})
 
-	const handleInputOnChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-		const value = e.target.value
-		setStepTwoDetails({ ...stepTwoDetails, [e.target.name]: value })
-	}
-
-	const submitDetails = () => {
-		setDetails({
-			...details,
-			product_type: stepTwoDetails.product_type,
-			weight: parseInt(stepTwoDetails.weight),
-			waiting_time: parseInt(stepTwoDetails.waiting_time) || undefined,
-			offer_price: parseInt(stepTwoDetails.offer_price),
-			description: stepTwoDetails.description,
-		})
+	const nextPage = () => {
 		router.push(`/jobs/add/3`, undefined, { shallow: true })
 	}
 
@@ -48,57 +31,11 @@ const JobAddStepTwo = (props: JobAddInterface) => {
 					percent={2 / 4}
 				/>
 			</FormHeader>
-			<FormInputContainer>
-				<InputComponent
-					name="product_type"
-					labelEN="Product Type"
-					labelTH="ชนิดสินค้า"
-					value={stepTwoDetails.product_type}
-					handleOnChange={handleInputOnChange}
-				/>
-				<InputComponent
-					name="weight"
-					labelEN="Product Weight"
-					labelTH="น้ำหนักสินค้า"
-					type="number"
-					description="1 ตัน = 1,000 กิโลกรัม"
-					classifier="ตัน"
-					value={stepTwoDetails.weight || ""}
-					handleOnChange={handleInputOnChange}
-				/>
-				<InputComponent
-					name="offer_price"
-					labelEN="Offer Price"
-					labelTH="ราคาเสนอ"
-					description={`ราคาที่เหมาะสมของงานนี้คือ ${offerCalculator(parseInt(stepTwoDetails.weight))}`}
-					type="number"
-					classifier="บาท"
-					value={stepTwoDetails.offer_price || ""}
-					handleOnChange={handleInputOnChange}
-				/>
-				<InputComponent
-					name="waiting_time"
-					labelEN="Waiting Time"
-					type="number"
-					labelTH="เวลารอขึ้น - ลงสินค้า"
-					classifier="ชั่วโมง"
-					required={false}
-					value={stepTwoDetails.waiting_time || ""}
-					handleOnChange={handleInputOnChange}
-				/>
-				<InputComponent
-					name="description"
-					labelEN="Description"
-					labelTH="คำอธิบายเพิ่มเติม"
-					required={false}
-					value={stepTwoDetails.description}
-					handleOnChange={handleInputOnChange}
-				/>
-				<FormActions>
-					<SecondaryButton onClick={() => router.push(`/jobs/add/1`)}>ย้อนกลับ</SecondaryButton>
-					<PrimaryButton onClick={submitDetails}>ส่วนถัดไป</PrimaryButton>
-				</FormActions>
-			</FormInputContainer>
+			<JobFormStepTwo />
+			<FormActionsCustom>
+				<SecondaryButton onClick={() => router.push(`/jobs/add/1`)}>ย้อนกลับ</SecondaryButton>
+				<PrimaryButton onClick={nextPage}>ส่วนถัดไป</PrimaryButton>
+			</FormActionsCustom>
 		</div>
 	)
 }
